refactor: 의존성 인터페이스 화
